Fusion: move visitors to a separate file to reuse them in loop fusion
